update sentry and add exception
